+++ dev/com.ibm.ws.security.openidconnect.clients.common/src/com/ibm/ws/security/openidconnect/clients/common/OidcClientConfig.java
@@ -11,7 +11,6 @@
 package com.ibm.ws.security.openidconnect.clients.common;
 
 import com.ibm.ws.security.common.structures.SingleTableCache;
-import com.ibm.ws.webcontainer.security.OidcSessionCache;
 
 public interface OidcClientConfig extends ConvergedClientConfig {
